+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # development version
 
+- Add Wallender 2021 piperaquine PK, malaria-hazard and QTc models ([doi:10.1038/s41467-021-27051-8](https://doi.org/10.1038/s41467-021-27051-8)) -- Ugandan children 2 to 36 months of age receiving malaria chemoprevention.
+
 - Add Ousey 2026 plozasiran ([doi:10.1002/jcph.70190](https://doi.org/10.1002/jcph.70190)) -- adults with familial chylomicronemia syndrome.
 
 - Add Ravva 2010 varenicline exposure-response models ([doi:10.1038/clpt.2009.282](https://doi.org/10.1038/clpt.2009.282)) -- adult cigarette smokers in smoking-cessation trials.